- fix issue with button background color
