+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/VisualControlsPage.xaml.cs
@@ -13,14 +13,10 @@ namespace Xamarin.Forms.Controls.Issues
 {
 	public partial class VisualControlsPage : TestShell
 	{
-		bool isVisible = false;
-		double percentage = 0.0;
-
 		public VisualControlsPage()
 		{
 #if APP
 			InitializeComponent();
-			Device.BeginInvokeOnMainThread(OnAppearing);
 #endif
 		}
 
@@ -28,47 +24,6 @@ namespace Xamarin.Forms.Controls.Issues
 		{
 			BindingContext = this;
 		}
-
-		public double PercentageCounter
-		{
-			get { return percentage; }
-			set
-			{
-				percentage = value;
-				OnPropertyChanged();
-				OnPropertyChanged(nameof(Counter));
-			}
-		}
-
-		public double Counter => percentage * 10;
-
-		protected override void OnAppearing()
-		{
-			isVisible = true;
-
-			base.OnAppearing();
-
-			Device.StartTimer(TimeSpan.FromSeconds(1), () =>
-			{
-				var progress = PercentageCounter + 0.1;
-				if (progress > 1)
-					progress = 0;
-
-				PercentageCounter = progress;
-
-				return isVisible;
-			});
-		}
-
-		protected override void OnDisappearing()
-		{
-			isVisible = false;
-
-			base.OnDisappearing();
-		}
-
-
-
 
 		[Preserve(AllMembers = true)]
 		[Issue(IssueTracker.Github, 4435, "Visual Gallery Loads",
